pyproject.toml
[project]
name = "discrete-egm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flow-matching",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "latex>=0.7.0",
    "matplotlib>=3.10.1",
    "numpy<2.0",
    "pot>=0.9.5",
    "seaborn>=0.13.2",
    "torch==1.13.1",
    "torchvision>=0.14.1",
    "tqdm>=4.67.1",
    "wandb>=0.19.8",
]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
flow-matching = { git = "https://github.com/facebookresearch/flow_matching.git" }
